droped react-media in uath pages, coz it throws error about stae in stateless component when component dont mount at all, rewrite uath pages
