+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,54 +1,18 @@
-import { useFormik } from 'formik';
-import Media from 'react-media';
+import { useMediaQuery } from '@mui/material';
+
 import AuthMedia from '../../components/Auth/AuthMedia';
 import AuthPageWrapper from '../../components/Auth/AuthPageWrapper';
-import AuthFormWrapper from '../../components/Auth/AuthFormWrapper';
-import AuthForm from '../../components/Auth/AuthForm';
-import {
-	EmailInputWithFormik,
-	PasswordInputWithFormik
-} from '../../components/Inputs';
+import LoginForm from '../../components/Auth/LoginForm';
 
 import styles from './LoginPage.module.css';
 
-import { loginFormValidationSchema } from '../../utils';
-
-const initialValues = {
-	email: '',
-	password: ''
-};
 export default function LoginPage() {
-	const formik = useFormik({
-		initialValues,
-		validationSchema: loginFormValidationSchema,
-		onSubmit: (values) => console.log(values)
-	});
+	const matches = useMediaQuery('(min-width:768px)');
 
 	return (
 		<AuthPageWrapper>
-			<Media
-				query="(min-width: 768px)"
-				render={() => <AuthMedia className={styles.bgImage} />}
-			/>
-			<AuthFormWrapper>
-				<AuthForm
-					formik={formik}
-					primaryBtnText="login"
-					secondaryBtnText="register"
-					navigateTo="/register"
-				>
-					<EmailInputWithFormik formik={formik} autoFocus />
-					<PasswordInputWithFormik
-						id="password"
-						name="password"
-						onChange={formik.handleChange}
-						value={formik.values.password}
-						error={formik.touched.password && Boolean(formik.errors.password)}
-						helperText={formik.touched.password && formik.errors.password}
-						placeholder="Password"
-					/>
-				</AuthForm>
-			</AuthFormWrapper>
+			{matches && <AuthMedia className={styles.bgImage} />}
+			<LoginForm />
 		</AuthPageWrapper>
 	);
 }
